backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
motor
python-jose[cryptography]
passlib[bcrypt]
pymongo
python-dotenv
email-validator
python-multipart
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fpdf2
qrcode[pil]==7.4.2
emails
Jinja2
reportlab==4.4.3
google-generativeai